[RENAME] 바텀시트 components/layout 폴더로 옮김
